% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R, R/s3_methods.R
\name{coerce_fpplateaus_data_frame}
\alias{coerce_fpplateaus_data_frame}
\alias{as_fpplateaus_data_frame}
\alias{as_fpplateaus_data_frame.default}
\alias{as_fpplateaus_data_frame.data.frame}
\alias{as_fpplateaus_data_frame.fpplateaus_data_frame}
\alias{is_demog_change_component_df}
\alias{as.data.frame.fpplateaus_data_frame}
\title{Coerce to a \code{fpplateaus_data_frame}}
\usage{
as_fpplateaus_data_frame(x, ...)

\method{as_fpplateaus_data_frame}{default}(x, ...)

\method{as_fpplateaus_data_frame}{data.frame}(x)

\method{as_fpplateaus_data_frame}{fpplateaus_data_frame}(x)

is_demog_change_component_df(x)

\method{as.data.frame}{fpplateaus_data_frame}(x, ...)
}
\arguments{
\item{x}{An object to coerce or check.}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A coerced object in the case of the \code{as_...}
functions; a logical for the \code{is_...} functions.
}
\description{
These functions coerce an object to a \code{fpplateaus_data_frame}
if possible, or check if it is one. The method for
\code{as.data.frame} will strip \code{x} of all its special
attributes and set the class to \code{"data.frame"}.
}
\author{
Mark Wheldon
}
