% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{count_n_plateau_countries}
\alias{count_n_plateau_countries}
\title{Make table reporting number of countries with at least one plateau year}
\usage{
count_n_plateau_countries(x)
}
\arguments{
\item{x}{Result of \code{\link{make_n_stall_years_list}}.}
}
\value{
Integer.
}
\description{
Make table reporting number of countries with at least one plateau year
}
\author{
Mark Wheldon
}
