% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{fpplateaus_data_frame}
\alias{fpplateaus_data_frame}
\alias{is_demog_change_component_df}
\title{Constructor for class \code{fpplateaus_data_frame}}
\usage{
fpplateaus_data_frame(
  x,
  differences = double(),
  change_condition_as_proportion = double(),
  filter_width = double(),
  denominator_count_filename = character(),
  stall_probability_thresholds = double(),
  CP_range_condition_min = double(),
  CP_range_condition_max = double(),
  MDMM_range_condition_min = double(),
  MDMM_range_condition_max = double(),
  min_stall_length = double()
)

is_demog_change_component_df(x)
}
\value{
An object of class \code{fpplateaus_data_frame}.
}
\description{

}
\details{

}
\author{
Mark Wheldon
}
