% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPEMglobal_aux_functions.R
\name{get_country_traj_muw}
\alias{get_country_traj_muw}
\title{Load and return country-level trajectories for married and unmarried women}
\usage{
get_country_traj_muw(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  iso_code,
  verbose = FALSE
)
}
\arguments{
\item{iso_code}{Numeric ISO code of country to get trajectories
for.}
}
\value{
The loaded country trajectory object.
}
\description{
This function \code{\link{load}}s the MCMC sample from the
posterior sample for married or unmarried for a single country and returns it as an R
object. Trajectories are loaded from \file{.rda} files found in
the subdirectory \file{countrytrajectories} of the results
directory (see below). The filename for the given country is
determined by reference to an index which must be in the file
\code{file.path(output_dir, "iso.Ptp3s.key.csv")}.
}
\details{
Country trajectories are 3D matrices:
\preformatted{str(...)
num [1:61, 1:3, 1:13344] 0.0622 0.0725 0.0879 0.0633 0.1078 ...
- attr(*, "dimnames")=List of 3
 ..$ : chr [1:61] "1970.5" "1971.5" "1972.5" "1973.5" ...
 ..$ : chr [1:3] "Traditional" "Modern" "Unmet"
 ..$ : NULL}

\subsection{Married/unmarried vs all women trajectories}{
The country trajectories for all women are stored differently. Do
not use this function for all women country trajectories. Use
\code{\link{get_country_traj_aw}} instead.

The country trajectories for all women are \emph{counts}. Those
for married and unmarried are \emph{proportions}.}
}
\section{Specifying results directory}{

See the section in \code{\link{get_csv_res}}.
}

\seealso{
Other trajectory functions: 
\code{\link{get_country_traj_aw}()}
}
\author{
Mark Wheldon
}
\concept{trajectory functions}
