% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPEMglobal_aux_functions.R
\name{get_csv_denominators}
\alias{get_csv_denominators}
\title{Get denominator counts from csv files}
\usage{
get_csv_denominators(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  filename = NULL,
  marital_group = c("default", "married", "unmarried", "all women"),
  add_marital_group = length(marital_group) > 1L,
  age_group = NULL,
  add_age_group = TRUE,
  clean_col_names = TRUE,
  units = c("units", "thousands"),
  years_as_midyear = TRUE,
  table_format = c("long", "raw"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{filename}{Name of file with the counts (including
extension). If \code{NULL}, this will be inferred from the
meta data (see \code{\link{get_global_mcmc_args}}.}

\item{marital_group}{Marital group to load denominators for. Can
be more than one; see \dQuote{Details}. The default is the
length 1 vector \code{"default"}.}

\item{add_marital_group}{Logical. Should a column
\dQuote{\code{marital_group}} be added to the output to
indicate the marital group? Such a column is always added if
\code{marital_group} has more than one element.}

\item{age_group}{Age group of the counts for the output data
frame. If the column names are of the form
'U/MW_[aabb]_year' this is ignored and the age group is
taken from the column names.}

\item{units}{Units in which to return the counts.}
}
\value{
A \code{\link[tibble]{tibble}} with the denominators; see
\dQuote{Details}.
}
\description{
Reads the '.csv' file containing the married and unmarried
denominator counts used in the run. These are in units of the
individual.

The data frame returned has columns \code{"iso"}, \code{"name"},
\code{"year"}, \code{"count"}. The \code{"year"} column will have
years in \dQuote{1st Jan} format (i.e., whole number years, 1970,
1971 etc.) because this is the format in which they are stored in
the \file{csv} files. This differs from
\code{\link{get_used_denominators}}. There is no option to return
\dQuote{mid-year} format years from \code{get_csv_denominators}.
}
\details{
One or more marital groups can be requested via argument
\code{marital_group}. If more than one, a column
\dQuote{\code{marital_group}} will be present in the output to
identify records accordingly (regardless the value of
\code{add_marital_group}). Value \code{"default"} will read the
the meta data (see \code{\link{get_global_mcmc_args}}) and ensure
the marital group matching that of the output is included. If
\code{"all women"} is included, values will be constructed by
summing the married and unmarried denominators.

If \code{table_format} is \code{"long"}, the counts are in a
single column called \dQuote{\code{value}}, with the necessary
identifying columns added. Otherwise the result is in the same
format as the original counts files, which have one column per
year.

The counts appear in the \file{.csv} files in units of 1 and, by
default, are returned as such (\code{units = "unit"}). Use the
\code{units} argument to return counts in multiples of 1000.
}
\section{Technical Note}{

The poulation denominators are stored in \file{.csv} files in the
output directory. These are read in using an unexported function
from \pkg{FPEMglobal}, which is a thin wrapper to \code{read.csv}.
}

\author{
Mark Wheldon
}
\concept{input_data_functions}
