% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{get_fp_plateau_countries}
\alias{get_fp_plateau_countries}
\title{Filter results of plateau analysis}
\usage{
get_fp_plateau_countries(
  x,
  stall_probability = NULL,
  indicator = NULL,
  year_lim = NULL,
  .filter = TRUE
)
}
\arguments{
\item{x}{Data frame loaded from e.g., 'wra_all_res.rda'.}

\item{stall_probability}{Probability threshold for a stall.}

\item{indicator}{FP indicator to keep.}

\item{year_lim}{Limits of time frame to keep. Leave as \code{NULL} (default) to keep all years.}

\item{.filter}{Logical; actually filter out non-stall countries?}
}
\value{
Filtered version of \code{x}.
}
\description{
Filters 'xxx_all_res_rda' data frame, returning only countries
with at least one plateau period for the indicator and time period
specified. Adds \code{FP_plateau} (logical) column indicating
whether the respective year is in a plateau period.
}
\author{
Mark Wheldon
}
