% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPEMglobal_aux_functions.R
\name{get_used_unpd_regions}
\alias{get_used_unpd_regions}
\title{Get UNPD aggregate country classifications}
\usage{
get_used_unpd_regions(
  run_name = NULL,
  output_dir = NULL,
  root_dir = NULL,
  clean_col_names = TRUE,
  UNlocations_names = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{UNlocations_names}{Logical; should
\code{\link{match_UNlocations}(..., return_names = "UNlocations")} be
run to standardize country and area names?}
}
\value{
A \code{\link[tibble]{tibble}} with the requested results.
}
\description{
Loads the country classifications actually used in a model
run. The files are read using \code{\link[readr]{read_csv}}.
}
\seealso{
Other countries, regions and aggregates functions: 
\code{\link{get_195_countries}()}

Other country_aggregates: 
\code{\link{get_195_countries}()}
}
\author{
Mark Wheldon
}
\concept{countries, regions and aggregates functions}
\concept{country_aggregates}
