% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isos_test_countries}
\alias{isos_test_countries}
\title{Country ISO codes using in testing}
\format{
\subsection{\code{isos_test_countries}}{

A numeric vector.
}
}
\source{
See \url{https://www.un.org/development/desa/pd/data/family-planning-indicators} and \url{https://github.com/FPcounts/FPEMglobal}
}
\usage{
isos_test_countries
}
\description{
Vector of the numeric codes of the countries included
in the test data files \code{link{sample_trajectories_all_women}}
and \code{link{sample_trajectories_married_women}}, and the sample
output files \code{\link{sample_output_wra_all_res_df}} and
\code{\link{sample_output_mwra_all_res_df}}.
}
\keyword{datasets}
