% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_results.R
\name{load_all_plateaus}
\alias{load_all_plateaus}
\title{Simple wrapper to load all results for a given rate condition threshold}
\usage{
load_all_plateaus(path_list)
}
\arguments{
\item{path_list}{List which must have names \dQuote{rate_01},
\dQuote{rate_03}, \dQuote{rate_05}.}

\item{rate_cond_thold}{Rate condition threshold to load.}
}
\value{
A list of data frames.
}
\description{
Loads results for a single method (e.g., linear smoothing) and a
single rate condition threshold. Returns a list of data frames,
one for each combination of indicator, marital group and
probability condition threshold.
}
\author{
Mark Wheldon
}
