% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{make_all_results_list}
\alias{make_all_results_list}
\title{Make list of all results objects}
\usage{
make_all_results_list(
  x,
  stall_probability = 0.8,
  indicator = "Modern",
  year_lim = c(1980.5, 2019.5)
)
}
\arguments{
\item{x}{\code{\link{fpplateaus_data_frame}} loaded from e.g., 'wra_all_res.rda'.}

\item{stall_probability}{Probability threshold for a stall.}

\item{indicator}{FP indicator to keep.}

\item{year_lim}{Limits of time frame to keep. Leave as \code{NULL}
(default) to keep all years.}
}
\value{
List of all summaries.
}
\description{
Wrapper function that calls all summaries and returns them in a
list.
}
\author{
Mark Wheldon
}
