% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{make_fp_plateau_only_results_table}
\alias{make_fp_plateau_only_results_table}
\title{Make summary table of FP plateaus}
\usage{
make_fp_plateau_only_results_table(x, require_level_condition = TRUE)
}
\arguments{
\item{x}{Output of \code{\link{make_main_results_df}}.}

\item{require_level_condition}{Logical; keep only periods which
satisfy the level condition?}
}
\value{
Data frame, possibly with zero rows.
}
\description{
Creates a data frame that can be printed to summarize plateaus by
country and type. Only retains countries with a
plateau. Additionally, if \code{require_level_condition = TRUE}
(defualt), only periods that satisfy the level condition are
retained. This could result in no matches, in which case a data
frame with zero rows is returned.
}
\details{
Compare with \code{\link{make_main_results_table}} which
includes TFR stalls in the blocking of time periods.
}
\seealso{
Other Main results tables: 
\code{\link{make_main_results_table}()}
}
\author{
Mark Wheldon
}
\concept{Main results tables}
