% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{make_main_results_df}
\alias{make_main_results_df}
\title{Make data frame of main results for plateau analysis}
\usage{
make_main_results_df(
  x,
  stall_probability,
  indicator,
  year_lim,
  CP_plateau_type_break = 0.5,
  MDMM_plateau_type_break = 0.75
)
}
\arguments{
\item{x}{Data frame loaded from e.g., 'wra_all_res.rda'.}

\item{stall_probability, indicator, year_lim}{See
\code{\link{get_fp_plateau_countries}}.}

\item{CP_plateau_type_break, MDMM_plateau_type_break}{Breakpoint in
indicator range to use to generate two plateau types.}
}
\value{
Data frame.
}
\description{
Construct new columns counting number of stall years,
etc. \emph{Note:} This function can work on the un-filtered
results, i.e., countries with neither a stall nor a plateau are
included in the input.
}
\author{
Mark Wheldon
}
