% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{make_n_stall_years_list}
\alias{make_n_stall_years_list}
\title{Create a list with summaries of stalls at various geographic aggregations}
\usage{
make_n_stall_years_list(x)
}
\arguments{
\item{x}{Output of \code{\link{make_main_results_df}}.}
}
\value{
List of data frames.
}
\description{
Makes \code{n_stall_years_...} data frames at world, region, and country level.
}
\author{
Mark Wheldon
}
