% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{make_tbl_country_n_plateau_years}
\alias{make_tbl_country_n_plateau_years}
\title{Make table reporting number of years of plateaus by country}
\usage{
make_tbl_country_n_plateau_years(x, rename = NULL)
}
\arguments{
\item{x}{Result of \code{\link{make_all_results_list}}.}

\item{rename}{Character string; new name for the \dQuote{value} column.}
}
\value{
Data frame that can be used as a table.
}
\description{
Make table reporting number of years of plateaus by country
}
\author{
Mark Wheldon
}
