% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{new_fpplateaus_data_frame}
\alias{new_fpplateaus_data_frame}
\title{Low-level constructor for class \code{fpplateaus_data_frame}.}
\usage{
new_fpplateaus_data_frame(
  x = data.frame(),
  marital_group = character(),
  differences = double(),
  change_condition_as_proportion = double(),
  filter_width = double(),
  denominator_count_filename = character(),
  stall_probability_thresholds = double(),
  CP_range_condition_min = double(),
  CP_range_condition_max = double(),
  MDMM_range_condition_min = double(),
  MDMM_range_condition_max = double(),
  min_stall_length = double(),
  FP_plateau_types = character(),
  Level_condition_variant = character(),
  ...,
  class = character()
)
}
\description{
Creates an object of class \code{fpplateaus_data_frame}. Minimal
checks are done; for interactive use see
\code{\link{fpplateaus_data_frame}}.

This function is not exported. The user-level constructor is
\code{\link{fpplateaus_data_frame}}.
}
\seealso{
fpplateaus_data_frame
}
\author{
Mark Wheldon
}
