% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{period_compare_plot}
\alias{period_compare_plot}
\title{Make plot comparing FP and TFR plateau periods}
\usage{
period_compare_plot(
  x,
  keep_only_stalls_c = TRUE,
  by_FP_plateau_type = TRUE,
  CP_abbrev = "MCP",
  use_ggpattern = FALSE
)
}
\arguments{
\item{x}{Result of \code{\link{make_main_results_df}}.}

\item{keep_only_stalls_c}{Logical; only plot countries with a stall, plateau or both.}

\item{by_FP_plateau_type}{Logical; separate the FP plateaus by type.}

\item{CP_abbrev}{Label for FP plateaus in the legend.}

\item{patterns}{Logical; use \pkg{ggpattern} functions to fill in the bars.}
}
\value{
\code{\link{ggplot2}} plot (invisibly).
}
\description{
This is the horizontal bar chart showing years with CP plateaus and FP stalls.
}
\details{
Note that the input is an already filtered data frame, so if you
want the plot for, e.g., married/in-union women you need to create
the appropriate input data set.
}
\author{
Mark Wheldon
}
