% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{plateau_compare_def_plot}
\alias{plateau_compare_def_plot}
\title{Compare plateaus plot}
\usage{
plateau_compare_def_plot(
  c_code,
  res_05_df,
  res_03_df,
  res_01_df,
  CP_abbrev = "MCP",
  probability_scale = c("percent", "prop"),
  line_colour = "black",
  fill_legend_title = "Plateau Type",
  linetype_legend_title = "Rate Condition Threshold",
  Level_condition_variant
)
}
\arguments{
\item{c_code}{Numeric country code}

\item{res_05_df, res_03_df, res_01_df}{Results to plot and compare;
results of \code{\link{load_all_plateaus}}. Currently, \emph{must} be
for rate condition thresholds 0.5, 0.3, and 0.1 percentage
points.}

\item{CP_abbrev}{Label for FP plateaus in the legend.}

\item{probability_scale}{Percent or proportion.}

\item{line_colour}{Line colour}

\item{fill_legend_title}{Title for fill legend.}

\item{linetype_legend_title}{Title for linetype legend.}
}
\value{
A ggplot.
}
\description{
Wrapper for \code{\link{stall_plot}} that shows stall
probabilities for different rate condition thresholds.
}
\author{
Mark Wheldon
}
