% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{plateau_ts_plot}
\alias{plateau_ts_plot}
\title{Side-by-side time series plots}
\usage{
plateau_ts_plot(
  c_code,
  res_df,
  res_03_df = NULL,
  res_01_df = NULL,
  CP_abbrev = "MCP",
  CP_not_in_range_abbrev = CP_abbrev,
  yvar_fp_plot = c("Modern_median", "stall_prob"),
  probability_scale = c("percent", "prop"),
  add_TFR_stalls = TRUE,
  legend_title = "Plateau Type",
  linetype_legend_title = "Rate Condition Threshold",
  stall_prob_line_colour = "black"
)
}
\arguments{
\item{c_code}{Numeric country code}

\item{CP_abbrev}{Label for FP plateaus in the legend.}

\item{probability_scale}{Percent or proportion.}

\item{linetype_legend_title}{Title for linetype legend.}
}
\description{
Wrapper for \code{\link{stall_plot}} and \code{\link{plateau_compare_def_plot}} to produce side-by-side time
series plots of FP indicators and plateau probabilities.
}
