% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{plateau_ts_plot_multiple}
\alias{plateau_ts_plot_multiple}
\title{Side-by-side time series plots}
\usage{
plateau_ts_plot_multiple(
  res_df,
  indicator_col_title = "Modern Contraceptive Prevalence",
  plateau_prob_col_title = "Plateau Probability",
  ...
)
}
\arguments{
\item{indicator_col_title}{Text string; title for the indicator column of plots.}

\item{plateau_prob_col_title}{Text string; title for the plateau probability column of plots.}

\item{...}{Passed to \code{\link{plateau_ts_plot}}.}
}
\value{
Called for side effect of printing a ggplot.
}
\description{
Wrapper for \code{\link{plateau_ts_plot}} to produce plots for
multiple countries in a single (multi-page) plot. This should be
wrapped by \code{\link{pdf}} ... \code{\link{dev.off()}} to have
output saved to a PDF file.
}
