% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_output_wra_all_res_df}
\alias{sample_output_wra_all_res_df}
\title{Sample output for all women}
\format{
\subsection{\code{sample_output_wra_all_res_df}}{

An \code{\link{fpplateaus_data_frame}} object.
}
}
\source{
See \url{https://www.un.org/development/desa/pd/data/family-planning-indicators} and \url{https://github.com/FPcounts/FPEMglobal}
}
\usage{
sample_output_wra_all_res_df
}
\description{
Ouptut generated by running \code{\link{make_all_results}} on
\code{link{sample_trajectories_all_women}}.
}
\section{Warning}{

These results are for demonstration purposes only. They are
\emph{not} suitable for use in analysis.
}

\seealso{
Other sample_outputs: 
\code{\link{sample_output_mwra_all_res_df}}
}
\concept{sample_outputs}
\keyword{datasets}
