% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_trajectories_all_women}
\alias{sample_trajectories_all_women}
\title{Sample of FP indicator trajectories for all women}
\format{
\subsection{\code{sample_trajectories_all_women}}{

A list of arrays. Each list element contains an array of
trajectories for a single country. Element are names are the ISO
numeric country codes. Only three countries are included: Nigeria
(ISO 566), Gambia (ISO 270), Zimbabwe (ISO 716). Each array has
\code{dim} \verbatim{\link{61, 6, 10}}. The first dimension is the
year, the second the indicator, and the third indexes
trajectory. There are ten trajectories in each array (i.e,. the
size of the third dimension is 10).
}
}
\source{
See \url{https://www.un.org/development/desa/pd/data/family-planning-indicators} and \url{https://github.com/FPcounts/FPEMglobal}
}
\usage{
sample_trajectories_all_women
}
\description{
A \emph{non-random}, \emph{non-representative} sample of indicator
\acronym{MCMC} trajectories for all women.
}
\section{Warning}{

These trajectory sets are for testing purposes only. They are
\emph{not} suitable for use in analysis.
}

\seealso{
Other sample_trajectories: 
\code{\link{sample_trajectories_married_women}}
}
\concept{sample_trajectories}
\keyword{datasets}
