% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{validate_fpplateaus_data_frame}
\alias{validate_fpplateaus_data_frame}
\title{Validate objects of class fpplateaus_data_frame.}
\usage{
validate_fpplateaus_data_frame(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
If valid, \code{x} is returned unchanged.
}
\description{
Validator function for objects of class \code{fpplateaus_data_frame}.
}
\author{
Mark Wheldon
}
