% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{fpplateaus_data_frame}
\alias{fpplateaus_data_frame}
\title{Constructor for class \code{fpplateaus_data_frame}}
\usage{
fpplateaus_data_frame(
  x,
  marital_group = attr(x, "marital_group"),
  differences = attr(x, "differences"),
  change_condition_as_proportion = attr(x, "change_condition_as_proportion"),
  filter_width = attr(x, "filter_width"),
  denominator_count_filename = attr(x, "denominator_count_filename"),
  stall_probability_thresholds = attr(x, "stall_probability_thresholds"),
  CP_range_condition_min = attr(x, "CP_range_condition_min"),
  CP_range_condition_max = attr(x, "CP_range_condition_max"),
  MDMM_range_condition_min = attr(x, "MDMM_range_condition_min"),
  MDMM_range_condition_max = attr(x, "MDMM_range_condition_max"),
  min_stall_length = attr(x, "min_stall_length"),
  FP_plateau_types = attr(x, "FP_plateau_types"),
  Level_condition_variant = attr(x, "Level_condition_variant")
)
}
\value{
An object of class \code{fpplateaus_data_frame}.
}
\description{
Creates objects of class \code{fpplateaus_data_frame}. Objects of
this class have class attribute
\code{c("fpplateaus_data_frame", "data_frame")}.

\code{is_fpplateaus_data_frame} is a simple wrapper for
\code{\link{inherits}}.
}
\author{
Mark Wheldon
}
