% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results_for_pub.R
\name{plateau_ts_plot}
\alias{plateau_ts_plot}
\title{Side-by-side time series plots}
\usage{
plateau_ts_plot(
  c_code,
  res_df,
  res_03_df = NULL,
  res_01_df = NULL,
  CP_abbrev = "MCP",
  CP_not_in_range_abbrev = CP_abbrev,
  yvar_fp_plot = c("Modern_median", "stall_prob"),
  probability_scale = c("percent", "prop"),
  Level_condition_variant,
  add_TFR_stalls = TRUE,
  legend_title = "Plateau Type",
  linetype_legend_title = "Rate Condition Threshold",
  stall_prob_line_colour = "black"
)
}
\description{
Wrapper for \code{\link{stall_plot}} to produce side-by-side time
series plots of FP indicators and plateau probabilities.
}
