% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{subset_replace}
\alias{subset_replace}
\alias{[.fpplateaus_data_frame}
\title{Extract or Replace Parts of a \code{fpplateaus_data_frame}}
\usage{
\method{[}{fpplateaus_data_frame}(x, i, j, drop)
}
\arguments{
\item{x}{data frame.}

\item{drop}{logical.  If \code{TRUE} the result is coerced to the
    lowest possible dimension.  The default is to drop if only one
    column is left, but \bold{not} to drop if only one row is left.}
}
\value{
An object of class inheriting from
\code{fpplateaus_data_frame} \emph{if} the result has at least
two columns and two rows.
}
\description{
These are methods for the subset operators and their companion
replacement functions for objects of class
\code{fpplateaus_data_frame}.
}
\seealso{
\code{\link{fpplateaus_data_frame}}.
}
\author{
Mark Wheldon
}
